[feat/bed-management] refactor common components
