+++ src/bundles/shared/components/index.js
@@ -1,0 +1,5 @@
+export * from './Datatable';
+export * from './Header';
+export * from './PatientMetadatum';
+export * from './TeamMetadatum';
+export * from './FilterList';
